+++ database/migrations/2025_07_08_091436_create_user_transaction_cycles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('user_transaction_cycles', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users');
-            $table->float('allocated_budget');
+            $table->integer('allocated_budget');
             $table->timestamp('active_from')->index('transaction_active_time_start_date');
             $table->timestamp('active_until')->index('transaction_active_time_end_date');
         });
